+++ cropper/src/main/java/com/canhub/cropper/CropWindowHandler.kt
@@ -160,11 +160,13 @@ class CropWindowHandler {
 
     /**
      * Determines which, if any, of the handles are pressed given the touch coordinates, the bounding
-     * box, and the touch radius.
+     * box, the touch radius, the crop shape and whether movement of the crop window is enabled.
      *
      * @param x the x-coordinate of the touch point
      * @param y the y-coordinate of the touch point
      * @param targetRadius the target radius in pixels
+     * @param cropShape the shape of the crop window
+     * @param isCenterMoveEnabled whether movement of the crop window by dragging center is enabled
      * @return the Handle that was pressed; null if no Handle was pressed
      */
     fun getMoveHandler(
@@ -172,12 +174,13 @@ class CropWindowHandler {
         y: Float,
         targetRadius: Float,
         cropShape: CropImageView.CropShape,
+        isCenterMoveEnabled: Boolean
     ): CropWindowMoveHandler? {
         val type: CropWindowMoveHandler.Type? = when (cropShape) {
-            RECTANGLE -> getRectanglePressedMoveType(x, y, targetRadius)
-            OVAL -> getOvalPressedMoveType(x, y)
-            RECTANGLE_VERTICAL_ONLY -> getRectangleVerticalOnlyPressedMoveType(x, y, targetRadius)
-            RECTANGLE_HORIZONTAL_ONLY -> getRectangleHorizontalOnlyPressedMoveType(x, y, targetRadius)
+            RECTANGLE -> getRectanglePressedMoveType(x, y, targetRadius, isCenterMoveEnabled)
+            OVAL -> getOvalPressedMoveType(x, y, isCenterMoveEnabled)
+            RECTANGLE_VERTICAL_ONLY -> getRectangleVerticalOnlyPressedMoveType(x, y, targetRadius, isCenterMoveEnabled)
+            RECTANGLE_HORIZONTAL_ONLY -> getRectangleHorizontalOnlyPressedMoveType(x, y, targetRadius, isCenterMoveEnabled)
         }
 
         return if (type != null) CropWindowMoveHandler(type, this, x, y) else null
@@ -192,12 +195,14 @@ class CropWindowHandler {
      * @param x the x-coordinate of the touch point
      * @param y the y-coordinate of the touch point
      * @param targetRadius the target radius in pixels
+     * @param isCenterMoveEnabled whether movement of the crop window by dragging center is enabled
      * @return the Handle that was pressed; null if no Handle was pressed
      */
     private fun getRectanglePressedMoveType(
         x: Float,
         y: Float,
-        targetRadius: Float
+        targetRadius: Float,
+        isCenterMoveEnabled: Boolean
     ): CropWindowMoveHandler.Type? {
 
         // Note: corner-handles take precedence, then side-handles, then center.
@@ -214,7 +219,8 @@ class CropWindowHandler {
             isInCornerTargetZone(x, y, mEdges.right, mEdges.bottom, targetRadius) -> {
                 CropWindowMoveHandler.Type.BOTTOM_RIGHT
             }
-            isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) &&
+            isCenterMoveEnabled &&
+                isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) &&
                 focusCenter() -> {
                 CropWindowMoveHandler.Type.CENTER
             }
@@ -230,7 +236,8 @@ class CropWindowHandler {
             isInVerticalTargetZone(x, y, mEdges.right, mEdges.top, mEdges.bottom, targetRadius) -> {
                 CropWindowMoveHandler.Type.RIGHT
             }
-            isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) &&
+            isCenterMoveEnabled &&
+                isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) &&
                 !focusCenter() -> {
                 CropWindowMoveHandler.Type.CENTER
             }
@@ -244,9 +251,14 @@ class CropWindowHandler {
      *
      * @param x the x-coordinate of the touch point
      * @param y the y-coordinate of the touch point
+     * @param isCenterMoveEnabled whether movement of the crop window by dragging center is enabled
      * @return the Handle that was pressed; null if no Handle was pressed
      */
-    private fun getOvalPressedMoveType(x: Float, y: Float): CropWindowMoveHandler.Type {
+    private fun getOvalPressedMoveType(
+        x: Float,
+        y: Float,
+        isCenterMoveEnabled: Boolean
+    ): CropWindowMoveHandler.Type? {
         /*
             Use a 6x6 grid system divided into 9 "handles", with the center the biggest region. While
             this is not perfect, it's a good quick-to-ship approach.
@@ -276,7 +288,9 @@ class CropWindowHandler {
             x < rightCenter -> {
                 when {
                     y < topCenter -> CropWindowMoveHandler.Type.TOP
-                    y < bottomCenter -> CropWindowMoveHandler.Type.CENTER
+                    y < bottomCenter -> if (isCenterMoveEnabled) {
+                        CropWindowMoveHandler.Type.CENTER
+                    } else null
                     else -> CropWindowMoveHandler.Type.BOTTOM
                 }
             }
@@ -297,12 +311,14 @@ class CropWindowHandler {
      * @param x the x-coordinate of the touch point
      * @param y the y-coordinate of the touch point
      * @param targetRadius the target radius in pixels
+     * @param isCenterMoveEnabled whether movement of the crop window by dragging center is enabled
      * @return the Handle that was pressed; null if no Handle was pressed
      */
     private fun getRectangleVerticalOnlyPressedMoveType(
         x: Float,
         y: Float,
-        targetRadius: Float
+        targetRadius: Float,
+        isCenterMoveEnabled: Boolean
     ): CropWindowMoveHandler.Type? {
 
         // Note: top and bottom handles take precedence, then center.
@@ -315,7 +331,8 @@ class CropWindowHandler {
             distance(x, y, mEdges.centerX(), mEdges.bottom) <= targetRadius -> {
                 CropWindowMoveHandler.Type.BOTTOM
             }
-            isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) -> {
+            isCenterMoveEnabled &&
+                isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) -> {
                 CropWindowMoveHandler.Type.CENTER
             }
             else -> null
@@ -329,12 +346,14 @@ class CropWindowHandler {
      * @param x the x-coordinate of the touch point
      * @param y the y-coordinate of the touch point
      * @param targetRadius the target radius in pixels
+     * @param isCenterMoveEnabled whether movement of the crop window by dragging center is enabled
      * @return the Handle that was pressed; null if no Handle was pressed
      */
     private fun getRectangleHorizontalOnlyPressedMoveType(
         x: Float,
         y: Float,
-        targetRadius: Float
+        targetRadius: Float,
+        isCenterMoveEnabled: Boolean
     ): CropWindowMoveHandler.Type? {
 
         // Note: left and right handles take precedence, then center.
@@ -347,7 +366,8 @@ class CropWindowHandler {
             distance(x, y, mEdges.right, mEdges.centerY()) <= targetRadius -> {
                 CropWindowMoveHandler.Type.RIGHT
             }
-            isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) -> {
+            isCenterMoveEnabled &&
+                isInCenterTargetZone(x, y, mEdges.left, mEdges.top, mEdges.right, mEdges.bottom) -> {
                 CropWindowMoveHandler.Type.CENTER
             }
             else -> null