+++ sample/src/main/java/com/canhub/cropper/sample/options_dialog/app/OptionsDialogBottomSheet.kt
@@ -168,6 +168,10 @@ internal class OptionsDialogBottomSheet : BottomSheetDialogFragment(), OptionsCo
             presenter.onMultiTouchSelect(isChecked)
         }
 
+        binding.centerMoveEnabled.toggle.setOnCheckedChangeListener { _, isChecked ->
+            presenter.onCenterMoveSelect(isChecked)
+        }
+
         binding.progressBar.toggle.setOnCheckedChangeListener { _, isChecked ->
             presenter.onProgressBarSelect(isChecked)
         }
@@ -210,6 +214,7 @@ internal class OptionsDialogBottomSheet : BottomSheetDialogFragment(), OptionsCo
 
         binding.autoZoom.toggle.isChecked = options.autoZoom
         binding.multiTouch.toggle.isChecked = options.multiTouch
+        binding.centerMoveEnabled.toggle.isChecked = options.centerMove
         binding.cropOverlay.toggle.isChecked = options.showCropOverlay
         binding.progressBar.toggle.isChecked = options.showProgressBar
         binding.flipHorizontal.toggle.isChecked = options.flipHorizontal